64-bit. Thanks to Umesh Shankar <ushankar@google.com>.

+++ ast/xmlhelp.h
@@ -9,7 +9,7 @@
 #include "str.h"         // string
 #include "srcloc.h"      // SourceLoc
 
-typedef unsigned int xmlUniqueId_t;
+typedef unsigned long xmlUniqueId_t;
 
 // manage identity canonicality; we now map addresses one to one to a
 // sequence number; this means that the ids should be canonical now
